Done with token logic on login and logout. Made useSubmit.ts hook for two forms.
TODO:
1. router.beforeEach

+++ src/services/api/setupInterceptors.ts
@@ -1,11 +1,11 @@
 import {axiosInstance} from "@/services/api/api";
 import router from "@/router";
 import TokenService from "@/services/token.service";
-import axios from "axios";
+import axios, {type AxiosRequestConfig} from "axios";
 
 export const setupInterceptors = () => {
     axiosInstance.interceptors.request.use(
-        (response) => {
+        (response: AxiosRequestConfig) => {
             const publicPages = ['/login', '/registration', '/login/restore'];
 
             if (!publicPages.includes(response.url!)) {
@@ -18,7 +18,7 @@ export const setupInterceptors = () => {
             return response;
         },
         (error: any) => {
-
+            return Promise.reject(error);
         }
     );
 
@@ -26,17 +26,20 @@ export const setupInterceptors = () => {
         (response) => {
             return response;
         },
-        async(error) => {
-            // dead access token
-            if (error.response.status === 401) {
+        async (error) => {
+
+            // Dead access token
+            if (error.response.status === 401 && TokenService.getAccessToken()) {
                 TokenService.removeAccessToken();
                 await router.replace({path: '/login'})
                 return Promise.reject('Access token expired');
             }
+
             // do not have rules / banned
-            if(error.response.status === 403) {
+            if (error.response.status === 403) {
 
             }
+            return Promise.reject(error);
         }
     )
 }
